pyproject.toml
[tool.poetry]
name = "noname_rpg_bot"
version = "0.1.0"
description = ""
authors = ["Artur <arturboyun@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.5"
aiogram = "3.0.0b5"
python-dotenv = "^0.21.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aerich = "^0.7.1"
click = "^8.1.3"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
black = "^22.8.0"


[tool.aerich]
tortoise_orm = "bot.database.db.TORTOISE_ORM_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
